ajusta regras de autorização de rotas: admin vê tudo

+++ src/user/user.controller.ts
@@ -1,39 +1,52 @@
-import { Controller, Get, Post, Put, Delete, Param, Body, UseGuards } from '@nestjs/common';
+import { Controller, Get, Post, Put, Delete, Param, Body, UseGuards, ForbiddenException } from '@nestjs/common';
 import { UserService } from './user.service';
-import { User } from './user.entity';
+import { User as UserEntity } from './user.entity';
 import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
+import { RolesGuard } from 'src/auth/roles.guard';
+import { Roles } from 'src/auth/roles.decorator';
+import { User } from 'src/auth/user.decorator';
 
 @Controller('users')
 export class UserController {
   constructor(private readonly userService: UserService) {}
 
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(JwtAuthGuard, RolesGuard)
   @Get()
-  findAll(): Promise<User[]> {
+  @Roles('admin')
+  findAll(@User() user: any): Promise<UserEntity[]> {
     return this.userService.findAll();
   }
 
   @UseGuards(JwtAuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: number): Promise<User> {
-    return this.userService.findOne(id);
+  @Roles('admin')
+  findOne(@Param('id') id: number, @User() user: any): Promise<UserEntity> {
+    if (user.role === 'admin' || user.userId == id) 
+      return this.userService.findOne(id);
+    throw new ForbiddenException('You do not have access to this resource');
   }
 
   @Post()
-  createUser(@Body() createUserDto: { username: string; password: string; role: 'admin' | 'student' | 'staff' }): Promise<User> {
+  createUser(@Body() createUserDto: { username: string; password: string; role: 'admin' | 'student' | 'staff' }): Promise<UserEntity> {
     const { username, password, role } = createUserDto;
     return this.userService.createUser(username, password, role);
   }
 
   @UseGuards(JwtAuthGuard)
   @Put(':id')
-  updateUser(@Param('id') id: number, @Body() updateData: Partial<User>): Promise<User> {
-    return this.userService.updateUser(id, updateData);
+  @Roles('admin')
+  updateUser(@Param('id') id: number, @User() user: any, @Body() updateData: Partial<UserEntity>): Promise<UserEntity> {
+    if (user.role === 'admin' || user.userId === id) 
+      return this.userService.updateUser(id, updateData);
+    throw new ForbiddenException('You do not have access to this resource');
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  deleteUser(@Param('id') id: number): Promise<void> {
-    return this.userService.deleteUser(id);
+  @Roles('admin')
+  deleteUser(@Param('id') id: number, @User() user: any): Promise<void> {
+    if (user.role === 'admin' || user.userId === id) 
+      return this.userService.deleteUser(id);
+    throw new ForbiddenException('You do not have access to this resource');
   }
 }
